removes updating the connection encode_context on trailer encoding
updating the hpack encoding is removed while trailers are
broken. Currently they can be sent in a different order than they
were encoded. A refactor to have the encoding not happen until
the sending is actually done is required to allow for the table
to be updated by trailer encoding.