Use proc_lib to spawn a helper
On window-changes, the h2_connection spawned off a helper to send as
much queued data as possible. Now use proc_lib for this so any crashes
in the spawned-off helper process don't go silently into the void.
Additionally, crash the h2 connection in case the spawned-off helper
crashes, since it is really an error, and ignoring it could lead to
other strange issues later.

Co-authored-by: mzeuner <max@zeuner.net>
Co-authored-by: JonathanArns <jonathan.arns@googlemail.com>